\name{heage}
\alias{heage}
\docType{data}
\title{
Northeast Thailand malnourishment data.
}
\description{
When screening for subclinical malnourishment standardized scores like the HE/AGE (based on an international reference population) are recommended. HE/AGE is computed as the difference of the childs height minus the median height in the reference population divided by the standadard deviation. Measurements of 708 preschool children in  northeast Thailand are contained in this data set.
}
\usage{data(heage)}
\format{
  A data frame with 708 observations on the following variable.
  \describe{
    \item{\code{value}}{a numeric vector}
  }
}
\references{
B{\"o}hning, D.  and Schlattmann, P. and Lindsay, B. (1992). \dQuote{Computer-Assisted Analysis of Mixtures (C.A.MAN): Statistical Algorithms}. \emph{Biometrics}. \bold{48}, 283--303.
}
\keyword{datasets}
