\name{hepab}
\alias{hepab}
\docType{data}
\title{
Hepatitis B data from Schlattmann and B{\"o}hning (1993).
}
\description{
This data set appears as an example for an application of mixture modelling in disease mapping. The data is on hepatitis B cases in Berlin 1989.
}
\usage{data(hepab)}
\format{
  A data frame with 23 observations on the following 2 variables.
  \describe{
    \item{\code{observations}}{a numeric vector}
    \item{\code{expected}}{a numeric vector}
  }
}
\references{
Schlattmann, P. and B{\"o}hning, D. (1993). \dQuote{Mixture models and disease mapping}. \emph{Statistics in Medicine}. \bold{12}, 1943--1950.
}

\keyword{datasets}
