\name{ret}
\alias{ret}
\docType{data}
\title{
Diet and betaplasma.
}
\description{
}
\usage{data(ret)}
\format{
  A data frame with 315 observations on the following 15 variables.
  \describe{
    \item{\code{id}}{a numeric vector}
    \item{\code{age}}{a numeric vector}
    \item{\code{sex}}{a factor with levels \code{F} \code{M}}
    \item{\code{smokestat}}{a factor with levels \code{Never} \code{Former} \code{Current}}
    \item{\code{bmi}}{a numeric vector}
    \item{\code{vituse}}{a factor with levels \code{Often} \code{Not often} \code{No}}
    \item{\code{calories}}{a numeric vector}
    \item{\code{fat}}{a numeric vector}
    \item{\code{fiber}}{a numeric vector}
    \item{\code{alcohol}}{a numeric vector}
    \item{\code{chol}}{a numeric vector}
    \item{\code{betadiet}}{a numeric vector}
    \item{\code{retdiet}}{a numeric vector}
    \item{\code{betaplasma}}{a numeric vector}
    \item{\code{retplasma}}{a numeric vector}
  }
}
\details{
See the example for the function \code{\link{mixcov}}.
}
\keyword{datasets}
