\name{vitA}
\docType{data}
\alias{vitA}
\title{Meta-analysis if vitamin A supplementation of childhood mortality}
\description{
Supplementation of vitamins is not only supposed to be beneficial for the prevention
of cancer as discussed in Sect. 2.1. Fawzi et al. (1993) study the effect of vitamin
A supplementation and childhood mortality in preschool children. They concluded
that vitamin A supplements are associated with a significant reduction in mortality
when given periodically to children at the community level.  All studies were
community randomized trials from South Asia or Southeast Asia, except the second
study, which was from northern Sudan. This data set contains the log relative risk and corresponding variance.
}
\usage{vitA}
\seealso{vitA2, vitamin}
\references{
Schlattmann, P.(2008) \emph{Medical Applications of Finite Mixture Models} Berlin: Springer
\cr
Fawzi,W.W., T. C. Chalmers, M. G. Herrera, and F. Mosteller (1993). Vitamin A supplementation
and child mortality. A meta-analysis. \emph{JAMA} 269(7), 898–903.

}
\keyword{datasets}
