\name{rs12363681}
\docType{data}
\alias{rs12363681}
\title{Gene calling}
\description{
This data set can be used to perform a cluster analyis of bivariate data.  \cr

This data set contains SNP of 3564 individuala 
}
\usage{data("rs12363681")}
\format{A data frame consisting of 15 data sets (rows) and 11 attributes (columns)}
\examples{
#Example
#Homogeneous Metaregression adjusting for study type and year of publication

data(rs12363681)


}
\references{
Schlattmann, P.(2009) \emph{Medical Applications of Finite Mixture Models.} Berlin: Springer.
}
\keyword{datasets}
